+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 [Unreleased](https://github.com/jshwi/jss/compare/v1.0.1...HEAD)
 ------------------------------------------------------------------------
+### Added
+- Adds versioning for posts
+
 ### Fixed
 - `Task` inherits from `_BaseModel`